requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
coveralls==1.7.0
Django==2.2
djangorestframework==3.9.2
docopt==0.6.2
gunicorn==19.9.0
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==7.0.0
pluggy==0.9.0
psycopg2-binary==2.8.2
py==1.8.0
pylint==2.3.1
pytest==4.4.1
pytest-cov==2.6.1
pytest-django==3.4.8
pytz==2019.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.1
